# openapi.yml
openapi: 3.0.0
info:
  title: Electron MCP REST API
  version: 1.0.0
  description: REST API for Electron MCP tools - 36 tools available
servers:
  - url: http://localhost:8101
    description: Local server
  - url: https://gcp-8101.cicy.de5.net
    description: Remote server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: token
security:
  - bearerAuth: []
paths:
  /rpc/tools:
    get:
      summary: List all available tools
      tags:
        - Tools
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of tools
          content:
            application/json:
              schema:
                type: object
                properties:
                  tools:
                    type: array
                    items:
                      type: object
                      properties:
                        name:
                          type: string
                        description:
                          type: string
...